TEMPORARY: Deprecate Str\join
Summary:
=== What is this gross function?
Currently, this API is inconsistent with the rest of the library, whose principle is that the "element being operated on" (in this case, the target Traversable) should be the first element. This temporary function will be used to correct that inconsistency.

=== Why are you disrupting my workflow?
We recognize that this will be disruptive in the short term and apologize for the inconvenience. However, it provides the benefit of predictability for all future users of the HSL.

=== But why now?
We are expecting to mark the HSL as stable soon, so this is our only chance to fix this.

=== Okay, fine... what's the game plan?
All uses of Str\join will be codemodded to `Str\join_args_switched`, and `Str\join` will be temporarily deprecated. After a short period, `Str\join` will be fixed, all callsites will be codemodded back, and this function will be removed.

=== Commands
- `hh --list-files | phps StrJoinSplitCodemod --stdin`
- `scripts/logger/create_typed_logger.php InstantArticlesTopicsLoggerConfig`
- `scripts/ui/asset gencode facebook_icons --merge`
- `./scripts/static_resources/no_more_sev_ones_please.pl`

drop-conflicts
bypass-lint
allow_many_files

Reviewed By: fredemmott

Differential Revision: D5945479

fbshipit-source-id: e5b07ebd74e8ccc46f1e164200e49f4d329b564c