Add 'C\all' comment to C\every for SEO
Summary:
Python has `any` and `all`:
https://docs.python.org/3/library/functions.html

JS has `some` and `every`:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/every

C++ has `any_of` and `all_of`:
http://www.cplusplus.com/reference/algorithm/any_of/
http://www.cplusplus.com/reference/algorithm/all_of/

I don't know what all the other languages do, but the point is that some folks may look at `C\any` and expect `C\all` to exist and that's not unreasonable.

To help guide them to the right function, I suggest we add a comment to `C\every` with the term `C\all` in it so folks can find this function more easily.

Reviewed By: fmoo, kmeht

Differential Revision: D9975183

fbshipit-source-id: 17dab35b4c425bbc33c176dba6f11330175bf70d