Fix Dict\fb\gen_map_with_key_limit_concurrency and friends to properly deal with failures
Summary:
Today if one of the concurrent workers throws an exception, it stops working, while the rest continue. This can be a problem if all but 1 fail, and this can greatly decrease the expected concurrentness. My fix is to save the exception, keep going on the worker, and then throw the first Exception (basically the same behavior as before).

In addition to this fix I also:
 - stabilized the result, so we will be a drop in replacement for Dict\gen_map (like the comment suggests)
 - no longer process in reverse order.

Reviewed By: kmeht

Differential Revision: D13791074

fbshipit-source-id: 0ccba827885af997b4cfa124ba74acd25038a6fb