Add time and space complexity to `C` library documentation
Summary:
This adds the complexities of each of the operations to the `C` library documentation. The complexities are what you would expect, with the exception of the `last` family of functions, which may be O(1) or O(n) depending on whether the input supports random access.

The complexities written down here assume that the cost of awaiting an `Awaitable` and getting the next element of a `Traversable` are O(1). I tried to be more explicit, but it's hard to convey this in a succinct manner, and doesn't seem to make the documentation much better.

Reviewed By: kmeht

Differential Revision: D13716705

fbshipit-source-id: e61472e7887154a1897b154fe5fb05852ed601c3