Optimize Math\from_base
Summary:
This is a performance optimization of Math\from_base, so that it converts directly to int from the source base instead of shifting bases and then using the decimal conversion.

There are a couple optimizations here; the most obvious is not converting to decimal, but less obvious is using a dict instead of Str\search_ci for getting the values of the digits.  Similarly, there is no Str\slice to get the subset of the alphabet in use; instead, the digit values are checked against the base for each digit (using just integer math).  Since negatives are not supported, overflow detection is quite simple.

I did *not* just shift to using intval because that handles illegal digits and overflows *very* differently.  In cases where we know that we have good inputs, intval is still about twice as fast as this implementation.

There are some additions to the test cases to cover more overflows, but otherwise there already was quite reasonable coverage of the input domain for Math\from_base.  However, this initial diff preserves the old implementation as Math\from_base_old, so that I could run benchmarks.  A second diff will follow shortly to remove the old implementation.

Reviewed By: pdillinger

Differential Revision: D14491063

fbshipit-source-id: e14b9f49ad77b2ea52a509be1a3bf65687ed260b