Optimize listDirectoryRecursively
Summary:
listDirectoryRecursively was painfully slow due to very repeated resolvePath and isDir calls. I realized that I'd be able to completely remove those calls if listDirectory would tell me the type of the items I was looking at, which it should be able to do anyway.

This ends up with listDirectoryByType.

Reviewed By: alexeyt

Differential Revision: D17063289

fbshipit-source-id: a6bb878a46c7306378d79df753dd8e3c92277fcf