Remove HH_FIXME[4275]
Summary:
If you remove the FIXME, you get this Hack error:

```
Typing[4275] Bad regex pattern; Missing delimiter(s).
   --> flib/core/hack/lib/regex/__tests__/RegexTest.php
 24 |     expect(() ==> $fn('foo', re"I am not a regular expression"))
    |                                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

I don't think there's any way to test that we throw on an invalid regex without making the typechecker complain about the invalid regex.

But I still think it's better to remove this test in order to remove the FIXME. Because Hack catches these invalid regexes for us, it should be a non-issue that we'd try to match against an invalid regex.

Reviewed By: kmeht

Differential Revision: D17124316

fbshipit-source-id: 0eb574d03fc668f21986227135133e082c777fd8