Make the `float` formatters take a `num` instead.
As @jjergus addressed in the case of [PlainSprintf](facebook/hhvm#8575) you can safely pass an int to these formatters. This is favorable over a `Str\format('Power level: %f', (float)$num)`, since the float cast will hide the typeerror in the case that `$num` would not be a number. At runtime it would silently convert `string(9) "Over 9000"` to `float(0.0)`.