remove HH\Lib\_Private\genw()
Summary:
1. This function is redundant (we have `genw()` internally, and `Asio\wrap()` in open-source)
2. It's not used in any open-source HSL code (only in `Dict\fb`)

So we really don't need to (and shouldn't) have it here.

It only has one callsite, which I replaced with `HH\Asio\wrap()` (I guess we could use `genw()` too, but it feels better to not use a FB-specific function from HSL code, even if it's the FB-specific part of HSL).

Reviewed By: kmeht

Differential Revision: D17698250

fbshipit-source-id: 40bdb04a9f36119d457a1bcd70cb610811f3ab3e