Split type parameters for contains and contains_key functions
Summary:
Use separate type parameters for the value and the value/key type of the traversable in `C\contains` and `C\contains_key`. This (a) simplifies life for the linter rule that fires on disjoint types for these two methods, as we now store the inferred type arguments in the TAST, and (b) will be used going forward when we introduce a special "NonDisjoint" attribute on type parameters.

Strictly speaking this is a breaking change, as more type arguments will be required in cases where they are specified explicitly.

Reviewed By: kmeht

Differential Revision: D17831267

fbshipit-source-id: 869743239b69b6ec9911030bd6c9692c64a4b5c8