Make Async\Semaphore free memory earlier
Summary:
Usage of async lambda + concurrent to build a dependency chain to track
condition notifiers has a number of issues:

- concurrent needs to hold all Awaitables in locals until the underlying
  AwaitAllWaitHandle is finished, so that it can properly throw exceptions;
  this means if there's an unfinished Awaitable deep in the tree, we will hold
  into all other Awaitables until it gets finished
- the above mentioned exception handling is dangerous and undesirable, as it
  will fail a ConditionWaitHandle if its dependency fails before it gets
  notified; this fortunately does not happen due to the order of operations,
  but it's better to avoid failing nodes of the notifiers tree
- so, the extra async lambda does not perform any useful work, it just creates
  unnecessary overhead and prevents memory from being freed

Let's fix this by directly using AwaitAllWaitHandle, like we do in Async\Poll.
The AwaitAllWaitHandle does exactly the work we need, has a memory effient
representation and drops references to its dependencies as soon as they are
finished.

Also, do not wrap `$this->activeGen` into an unnecessary closure when passed to
ConditionWaitHandle.

fixes hhvm/hsl#105

Reviewed By: fredemmott

Differential Revision: D19320896

fbshipit-source-id: ba8f982eaea40ecc8a11ed7af8413ebc92bd1fba