Backport Ref under v1.4.x's existing license
This file is being re-released under BSD+PATENTS to match the existing
license in this old release branch, to avoid including a license change in a .z
release.

Active development of the HSL continues under the MIT license, as indicated by
the LICENSE file in the master branch. The 1.4.x series was the last release
under the old license; the use of BSD+PATENTS here is only due to
backporting.

refs #108