Add C\findx() to the HSL (#128)
Summary:
This is part of the README.md file.

```
Find-like operations that can fail should return ?T;
a second function should be added with an x suffix
that uses an invariant to return T (e.g. C\first(), C\firstx())
```

There is a notable difference between `C\find(...) as nonnull` and `C\findx(...)` in the case where `null` would be a case where the `$value_predicate` returns true. There would be no way to know if you got a `null` returned, because there was a `null` in the `$traversable` or because nothing was found.

If `C\fb\findx()` is different / better than this, it may be promoted instead.
Pull Request resolved: https://github.com/hhvm/hsl/pull/128

Reviewed By: viratyosin

Differential Revision: D20349453

Pulled By: fredemmott

fbshipit-source-id: 5bd005132092c5624e9a1a43c216339b2891f48e