Mark regex_match reactive
Summary: These regex functions are pretty much reactive - the only nonreactive thing they do is the PHPWarningSuppressor. My original plan was to make a helper that extracted all the necessary error info from global state and returned all that data together, but as I used that function in the Regex namespace, I found that I was just doing the same thing the existing implementation did with extra steps. In my opinion, we should still throw if the user passes in an invalid regex, so I was just moving that logic to the callers of `_Private\regex_match()`.

Reviewed By: DavidSnider

Differential Revision: D21026626

fbshipit-source-id: 830b743846fd912991e166df09f784174cb1f24d