Permit using old release of hsl-experimental when testing
Summary:
dev-master of hsl-experimental currently requires a newer HHVM than
the main HSL does, due to the IO\ and OS\ work.

We either need to permit use of a release, or increase the minimum version of
HHVM for the HSL - I prefer the former :)

We generally use dev-master so that commits that change both end up passing on
Travis at roughly the same time.

Reviewed By: alexeyt

Differential Revision: D21044369

fbshipit-source-id: 41bef28f5b83a5fd9d0f6b7b7575f3ce64fb3234