Replace redundant generics with marker type
Summary:
This Runs the codemod in the previous diff.

This diff has some hack errors. They are cleaned up by the manual changes in the next diff and will be squashed before landing.

As stated in the previous diff, this will be squashed into that for landing. Separating for review.

drop-conflicts

Reviewed By: rodmk

Differential Revision: D21575682

fbshipit-source-id: 6ebd26c46f0396376e30a69b0f98cda71c48d578