Deal with cast-marked-array notices (part 1)
Summary:
Turning on "cast-marked-array" notices is a 5% regression, but that's just due to the head sites:

85% of notices come from the vec cast in Vec\gen_map.
80% of the remaining notices come from the dict cast in EntQLStructuredAssocQuery.
75% of the remaining notices come from the dict cast in Dict\gen_map.

Simply fixing these three sites brings the regression down to 0.03%. The remaining dict cast fixes should bring it down by yet another factor of ~3-4. In summary: this diff both fixes a bunch of notices and makes it safe to actually turn on the logging.

Reviewed By: DavidSnider

Differential Revision: D23983722

fbshipit-source-id: 3a0a2315f374a7cf7052ffa8411d266206406f72