Use Vec\cast_clear_legacy_array_mark in Vec\slice
Summary: We want HSL functions to produce new, unmarked arrays. Logging indicated that this `vec` cast was casting a marked PHP array. This can happen if `$container` is a non-empty, marked varray and the slice range covers the whole of the input; in this case, `array_slice` will end up returning the input.

Reviewed By: periodic1236, jthemphill

Differential Revision: D25893504

fbshipit-source-id: 3661c00c324c83caf0e041683325b23256511b2c