Make closure mangling deterministic
Summary:
Right now to ensure closure names are always unique,
we append a globally increasing counter to the name. This works, but
makes closure names non-deterministic (the counter is accessed by
multiple threads).

HackC already ensures that closure names are unique within a function,
and the enclosing function is part of the name.

The only time we can have collisions is if HHBBC clones the closure as
part of trait flattening. To ensure uniqueness in this context, append
the hash of the *destination* unit's filename, and the index of the
closure within the unit's class table. This combination is always
guaranteed to be globally unique. It is also deterministic.

There's some special logic in the runtime for defining a closure,
which assumed closure names were unique. Luckily, it can be easily
modified to allow for duplications. It needs to merely search the
class-list on the NamedEntity for the class with the same PreClass.

Differential Revision: D30354925

fbshipit-source-id: e48c97aca15eb9735ffe60c4265c911b69c236b4