Bring Regex hsl back in sync with internal copy
Summary: Migrating the Regex portion of the HSL to open source got stuck part way through, we forgot about it, and we then accidentally made changes to the internal copy without updating the open source copy. Bring changes from the internal copy over to the open source copy in order to get them back in sync so that we can resume the migration.

Reviewed By: DavidSnider

Differential Revision: D39677049

fbshipit-source-id: afcf439633884003b53d1b3fda7cec4d06046f8a