Improve PseudoRandom\float
Summary:
The previous implementation performed PHP division on integers that would first
check whether the modulo is zero, then perform integer division if it was zero,
followed by a side-exit, because it is rare that the PseudoRandom\int() returns
zero (the only value with modulo zero).

Instead, perform the float division.

Reviewed By: ricklavoie

Differential Revision: D49170059

fbshipit-source-id: ce60c05599b5be185c68d5e2a426ed3630d14b63