Revert D50089378: Multisect successfully blamed "D50089378: Remove hack modes from Hack tests" for test or build failures
Summary:
This diff is reverting D50089378
D50089378: Remove hack modes from Hack tests by viratyosin has been identified to be causing the following test or build failures:

Tests affected:
- [hphp/hack/test/facebook/write_symbol_info:regression_tests - cases/omega_analyser](https://www.internalfb.com/intern/test/844424983077690/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3260372
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D50206182

fbshipit-source-id: 3da6b0aca3e3e4be03884e58d43364408782f0e9