Remove TANY_MARKER type
Summary: The `TANY_MARKER<_>` type isn't used in www and is used only in `hhi` files to define the `MISSING_RETURN_TYPE` and `MISSING_PROP_TYPE` types. Defining these to be `dynamic` works fine with pessimisation, and with a very little strengthening of types in `hhi` files it works fine even in legacy mode. Note that `getOffset` and `getTimestamp` are now declared to return `int` in the interface `DateTimeInterface`. This is compatible with all implementing classes.

Differential Revision: D47090958

fbshipit-source-id: d134666dae6fd1897974f639e458b37ea62eec71