Add loose typing to debug_backtrace
Summary:
Sorry if this is not valid or the right way to do this, but I figured it'd be easiest to just try it out and get feedback. To motivate this change, I recently foolishly assumed `args` would always be present in a frame, which broke a bunch of stuff (D57301367). I think the demographic of folks who typically work closely with stacks would be familiar with this fickleness of frame data, but alas, as a noob I didn't have this intuition. It would've been helpful to have some typing to guide me in the right direction of using `idx` or foregoing the usage of the data entirely.

I based my typing off of [this docblock](https://www.internalfb.com/code/fbsource/[8eec5a24f012]/fbcode/hphp/runtime/ext/std/ext_std_errorfunc.php?lines=20-30) (which is almost certainly outdated, as seen by the omission of `metadata`) and [the code that generates the stack](https://fburl.com/code/0htc79f2). I opted for incredibly loose typing so as to not break anything. I know this is maybe *so* loose that it gives us almost no extra type safety, but I'm hoping that all the optional typing shows that nothing in the stack is guaranteed. If there is anything in the stack that you know is guaranteed, let me know and I can tighten it a bit.

Reviewed By: mheiber

Differential Revision: D57509960

fbshipit-source-id: 0c0fca48082f580255a125daed78918862d0ee8b