Read experimental feature status from config (step 1)
Summary:
Add a field to `.hhconfig` named `enable_experimental_stx_features`. It contains a json mapping experimental feature names to their status. For example,
`enable_experimental_stx_features = {"union_intersection_type_hints": "Unstable", "readonly": "Preview"}`. This is in preparation to move the configuration of experimental features out of the source code and into `.hhconfig` and `.hhvmconfig.hdf`.

Currently, we enforce that any feature specified in `enable_experimental_stx_features` has the same status as the hard-coded status, to ensure hh and hhvm stay in sync until we are ready to complete the migration.

There are 2 other configuration options:
- `consider_unspecified_experimental_features_released`: if an experimental feature is not listed in the config, should it default to `OngoingRelease`, or `Unstable` status. We set `hh_server` to `Unstable`, but allow all other tools (`hh_single_type_check`, `hh_parse`, `hackfmt`, etc.) to default to `OngoingRelease` since we expect them to have empty lists of experimental features as they don't read `.hhconfig`, and we don't need them to reject experimental features.
- `use_legacy_experimental_feature_config`: determines whether to use the new config-based mechanism, or the old hardcoded one. It is set to use the new mechanism for `hh_server`, and `hh_single_type_check`, and use the hardcoded mechanism for everything else. As we migrate, we will set this to the new mechanism for a wider range of entry points, and finally remove it altogether.

Reviewed By: CatherineGasnier

Differential Revision: D58869349

fbshipit-source-id: 43dd1d375d56540db709a8608eeaad6ea0c61311