Add Async\LowPri for bridging priority context states
Summary:
Adds a `LowPri` class to the `HH\Lib\Async` namespace, which provides a flexible way to run async lambdas in a low-priority context. This abstracts a newly created wait handle (PriorityBridgeWaitHandle) to achieve the following:

*   Start executing the async lambda upon calling `run()`, but only when there is no other regular priority work to be done.
*   The `run()` method initiates the low-priority action and returns an awaitable that resolves with the result without lifting the priority of the low-priority action.
*   The `prioritize()` method brings the low-priority action into the current context, lifting its priority.

## Basic Usage:
```
// Create a LowPri object
$low_pri_job = new HH\Lib\Async\LowPri<int>();

// Start the low-priority action by calling run() with an async lambda
$awaitable = $low_pri_job->run(async () ==> {
  // Low priority work here
  return 42;
});

// (Optionally) store the LowPri object for later prioritization
store($low_pri_job);

// Wait for the result, blocking but allowing work to execute in a low-priority context
$result = await $awaitable;

// (Optionally) Prioritize the job to lift its priority
$low_pri_job = load();
$low_pri_job->prioritize();
```

## Motivating Example
This API is particularly useful for product code that explicitly starts work in phases to achieve low latency on early work, but suffers from high tail latency due to inefficient use of spare CPU cycles in these early phases.

By changing these cases to start all of the tail work in a low priority mode, it can allow progress to be made when there's nothing else running (e.g. IO waits), and dramatically cut down tail latencies.

This could take the following form:
```
<<__EntryPoint>>
async function main(): Awaitable<void> {
  // Create tasks and store them
  $tasks = vec[
    tuple(new LowPri<int>(), async () ==> {
      echo "Start 1\n";
      await HH\Asio\usleep(3000000);
      echo "1 finished\n";
      return 1;
    }),
    tuple(new LowPri<int>(), async () ==> {
      echo "Start 2\n";
      await HH\Asio\usleep(2000000);
      echo "2 finished\n";
      return 2;
    }),
    tuple(new LowPri<int>(), async () ==> {
      echo "Start 3\n";
      await HH\Asio\usleep(1000000);
      echo "3 finished\n";
      return 3;
    }),
  ];

  // Start all tasks concurrently in a low priority context
  $running_tasks = vec[];
  foreach ($tasks as list($obj, $action)) {
    $running_tasks[] = tuple($obj, $obj->run($action));
  }

  // Prioritize each task, and wait for it to finish
  foreach ($running_tasks as list($obj, $task)) {
    $obj->prioritize();
    $res = await $task;
    echo "finished task=".$res."\n";
  }
}
```

In the example above, tasks 1, 2, and 3 will be started when `run()` is called. Then the tasks can be prioritized in the order that they're needed. The tasks in a low-pri mode (before `prioritize()` is called) will not steal CPU time from other non-low tasks. Specific to the example, everything will be done at the same time (tasks 2 & 3 are already done when task 1 finishes).

The net result of this is that all work can be started, and should not (meaningfully) impact latency of the early tasks.

Reviewed By: jano

Differential Revision: D75271660

fbshipit-source-id: aa15763833623d4e225f4df4e6fad22ee1cccf82